requirements.txt
pandas>=2.0.0
pyarrow>=12.0.0
openpyxl>=3.1.2
tiktoken>=0.5.1
openai>=1.12.0
pydantic>=2.5.0
python-dotenv>=1.0.0
pytest>=7.4.0
xlrd>=2.0.1 